refactor: criado hook e componente das bolinhas
