+++ pages/palestrantes.js
@@ -1,6 +1,6 @@
 import Meta from '../src/infra/seo/Meta';
 import styled from 'styled-components';
-import { useState, useEffect, useRef } from 'react';
+import { useState, useEffect } from 'react';
 
 // Components
 import PalestranteCard from '../src/components/features/speakers/PalestranteCard';
@@ -11,13 +11,11 @@ import saphira from '../services/saphira';
 import Image from 'next/image';
 
 import { eventDetails } from '../data/eventDetails'; 
+import useDotsGeneration from '../hooks/useDotsGeneration';
 
 const Palestrantes = () => {
     const [isLoading, setIsLoading] = useState(false)
     const [speakers, setSpeakers] = useState([])
-    // Ref e state para controlar o width disponível do componente das bolinhas
-    const speakerRef = useRef(null)
-    const [availableWidth, setAvailableWidth] = useState(0)
 
     const getSpeakers = async() => {
         setIsLoading(true)
@@ -37,31 +35,8 @@ const Palestrantes = () => {
         getSpeakers()
     }, [])
 
-    // Roda após receber os dados
-    useEffect(() => {
-        const scheduleComponent = speakerRef.current
-
-        if (!scheduleComponent) return;
-
-        // Função que pega o width disponível do componente das bolinhas
-        const updateAvailableWidth = () => {
-            const dotsWrapper = scheduleComponent.querySelector('.dots-wrapper');
-            if (!dotsWrapper) return;
-            
-            // Pega o width e atualiza o availableWidth se for diferente do valor atual
-            const nextWidth = dotsWrapper.getBoundingClientRect().width;
-            setAvailableWidth((currentWidth) => (currentWidth === nextWidth ? currentWidth : nextWidth));
-        }
-
-        updateAvailableWidth();
-        
-        // Listener para detectar mudanças no tamanho da tela
-        // E atualiza o availableWidth com o width disponível do componente das bolinhas
-        window.addEventListener('resize', updateAvailableWidth)
-
-        // Matando o listener quando o componente for desmontado
-        return () => window.removeEventListener('resize', updateAvailableWidth)
-    }, [speakers])
+    // Hook customizado para obter o width disponível do componente das bolinhas
+    const { componentRef, availableWidth } = useDotsGeneration(speakers);
 
     return (
         <PalestrantesContainer>
@@ -69,7 +44,7 @@ const Palestrantes = () => {
           description = {`Conheça os palestrantes da SSI ${eventDetails.year}! Referências em tecnologia, inovação e mercado de TI que compartilharão suas experiências com o público.`}
           keywords={`palestrantes SSI, especialistas em TI, convidados SSI ${eventDetails.year}, nomes da tecnologia, profissionais da tecnologia, lideranças em TI, conferencistas SSI, oradores evento TI`}
           />
-          <PalestrantesWrapper ref={speakerRef}>
+          <PalestrantesWrapper ref={componentRef}>
             <h1>Palestrantes</h1>
 
           {
