+++ src/components/features/speakers/PalestranteCard.js
@@ -1,7 +1,6 @@
 import { useEffect, useLayoutEffect, useRef, useState } from "react";
 import styled, { css } from "styled-components";
 import Image from "next/image";
-import PalestranteBolinhas from "./SpeakerDots";
 
 //IMAGES
 import speakersImages from "../../../../data/speakers";
@@ -11,6 +10,7 @@ import BackgroundPicture from "../../../../public/images/partnerships/photos/bg_
 
 import BadgeLecture from "../schedule/BadgeLecture";
 import formatActivityDateTime from "../../../../utils/formatActivityDateTime";
+import Dots from "../../ui/Dots";
 
 // Variável para controlar a exibição do badge de modo "Presencial"/"Online"
 const exibirBadgePresencial = false; 
@@ -152,7 +152,7 @@ const PalestranteCard = ({palestrante, availableWidth}) => {
                     </PalestranteLectureBody>
                 </PalestranteBody>
             </PalestranteWrapper>
-            <PalestranteBolinhas availableWidth={availableWidth}/>
+            <Dots dotSize={4} dotGap={10} availableWidth={availableWidth}/>
         </PalestranteContainer >
     )
 }
@@ -206,6 +206,15 @@ const PalestranteContainer = styled.div`
     flex-direction: column;
     flex: 1;
     position: relative;
+
+    // Cores do gradiente das bolinhas (Tem a mesma cor no dark e light mode)
+    .gradient-stop-1 {
+        stop-color: #666666;
+    }
+
+    .gradient-stop-2 {
+        stop-color: #666666;
+    }
 `;
 
 const PalestranteInfo = styled.div`
